requirements.txt
llama-index==0.10.1
groq==0.4.1
python-dotenv==1.0.1
